ref: Transfert des locales des modules dans leur dossier respectif
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@9081 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/hprimxml/ajax_file_to_echange.php
@@ -78,10 +78,12 @@ if (!$do_import) {
     $object = new $echg_hprim->object_class;
     $object->load($echg_hprim->object_id);
     $object->facture = 1;
-    $object->store();
+    $msg = $object->store();
 
     // Suppression du fichier sur le disque
-    unlink("$path/$_file");
+    if (!$msg) {
+      unlink("$path/$_file");
+    }
     
     $counter++;
